Fix `A copy of [segment] has been removed from the module tree but is still active!`
Segments are eager loaded, which also means they are auto reloaded in development. Since registering a segment was caching the segment instance, the instances were no longer valid after a code reload. This fixes the issue by registering the segment class and arguments instead of a concrete instance, and then calling `new` every time a segment needs to be accessed.