Fix `A copy of xxx has been removed from the module tree but is still active`
It turns out that the Audience segment registry was caching the class for each segment. When autoloading happens, the segment classes are reloaded, and the cached classes in the registry become invalid.

The solution is to store the name of the class in the registry, and `constantize` the class name every time a segment needs to be built.